drivers: irqsteer_1: fix compilation issue on ARM64 platform

+++ drivers/irqsteer_1/fsl_irqsteer.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 NXP
+ * Copyright 2024, 2026 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -78,7 +78,7 @@ void IRQSTEER_Init(int32_t instIdx)
         /* Mask all interrupts. */
         for (i = 0; i < data->regNum; i++)
         {
-            *(volatile uint32_t *)((uint32_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, i, data->regNum)) = 0U;
+            *(volatile uint32_t *)((uintptr_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, i, data->regNum)) = 0U;
         }
 
         for (i = 0; i < data->intGrpNum; i++)
@@ -135,7 +135,7 @@ void IRQSTEER_EnableInterrupt(int32_t instIdx, IRQn_Type irq)
     regIdx = IRQSTEER_GEN_REG_IDX(data->regNum, inputIdx);
     bitOffset = inputIdx % IRQSTEER_INT_SRC_REG_WIDTH;
 
-    *(volatile uint32_t *)((uint32_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, regIdx, data->regNum)) |= (1U << bitOffset);
+    *(volatile uint32_t *)((uintptr_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, regIdx, data->regNum)) |= (1U << bitOffset);
 }
 
 /*! @} */
@@ -161,7 +161,7 @@ void IRQSTEER_DisableInterrupt(int32_t instIdx, IRQn_Type irq)
     regIdx = IRQSTEER_GEN_REG_IDX(data->regNum, inputIdx);
     bitOffset = inputIdx % IRQSTEER_INT_SRC_REG_WIDTH;
 
-    *(volatile uint32_t *)((uint32_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, regIdx, data->regNum)) &= ~(1U << bitOffset);
+    *(volatile uint32_t *)((uintptr_t)data->infoPtr->reg + IRQSTEER_CHAN_MASK(data->infoPtr->irqChanIdx, regIdx, data->regNum)) &= ~(1U << bitOffset);
 }
 
 /**
@@ -266,7 +266,7 @@ IRQn_Type IRQSTEER_GetMasterNextInterrupt(int32_t instIdx, int32_t outputChanIdx
         regIdx = IRQSTEER_GetRegIdx(instIdx, outputChanIdx, i);
 
         /* Get register's value */
-        chanStatus = *(uint32_t *)((uint32_t)data->infoPtr->reg + IRQSTEER_CHAN_STATUS(data->infoPtr->irqChanIdx, regIdx, data->regNum));
+        chanStatus = *(uint32_t *)((uintptr_t)data->infoPtr->reg + IRQSTEER_CHAN_STATUS(data->infoPtr->irqChanIdx, regIdx, data->regNum));
 
         for (j = 0; j < IRQSTEER_INT_SRC_REG_WIDTH; j++)
         {
@@ -315,7 +315,7 @@ uint64_t IRQSTEER_GetMasterInterruptsStatus(int32_t instIdx, int32_t outputChanI
     for (i = 0; i < sliceNum; i++) {
         regIdx = IRQSTEER_GetRegIdx(instIdx, outputChanIdx, i);
 
-        chanStatus = *(uint32_t *)((uint32_t)data->infoPtr->reg + IRQSTEER_CHAN_STATUS(data->infoPtr->irqChanIdx, regIdx, data->regNum));
+        chanStatus = *(uint32_t *)((uintptr_t)data->infoPtr->reg + IRQSTEER_CHAN_STATUS(data->infoPtr->irqChanIdx, regIdx, data->regNum));
 
         interrupts |= ((uint64_t)chanStatus << (IRQSTEER_INT_SRC_REG_WIDTH * i));
     }
@@ -332,7 +332,7 @@ uint64_t IRQSTEER_GetMasterInterruptsStatus(int32_t instIdx, int32_t outputChanI
 void IRQSTEER_CommonIRQHandler(int32_t instIdx, int32_t outputChanIdx)
 {
     IRQn_Type intSource = NotAvail_IRQn;
-    int32_t isr = 0;
+    intptr_t isr = 0;
 
     intSource = IRQSTEER_GetMasterNextInterrupt(instIdx, outputChanIdx);
     if (NotAvail_IRQn != intSource)
